+++ e2e/tests/desktop/registered-shopper.spec.js
@@ -7,14 +7,6 @@
 
 const { test, expect } = require("@playwright/test");
 const config = require("../../config");
-const { 
-  addProductToCart,
-  registerShopper,
-  validateOrderHistory,
-  validateWishlist,
-  loginShopper,
-  navigateToPDPDesktop,
-} = require("../../scripts/pageHelpers");
 const {
   generateUserCredentials,
   getCreditCardExpiry,
@@ -22,18 +14,86 @@ const {
 
 const REGISTERED_USER_CREDENTIALS = generateUserCredentials();
 
-/**
- * Test that registered shoppers can add a product to cart and go through the entire checkout process,
- * validating that shopper is able to get to the order summary section,
- * and that order shows up in order history
- */
 test("Registered shopper can checkout items", async ({ page }) => {
-  // register and login user
-  await registerShopper({page, userCredentials: REGISTERED_USER_CREDENTIALS});
+  // Create Account and Sign In
+  await page.goto(config.RETAIL_APP_HOME + "/registration");
+
+  const registrationFormHeading = page.getByText(/Let's get started!/i);
+  await registrationFormHeading.waitFor();
+
+  await page
+    .locator("input#firstName")
+    .fill(REGISTERED_USER_CREDENTIALS.firstName);
+  await page
+    .locator("input#lastName")
+    .fill(REGISTERED_USER_CREDENTIALS.lastName);
+  await page.locator("input#email").fill(REGISTERED_USER_CREDENTIALS.email);
+  await page
+    .locator("input#password")
+    .fill(REGISTERED_USER_CREDENTIALS.password);
+
+  await page.getByRole("button", { name: /Create Account/i }).click();
+
+  await expect(
+    page.getByRole("heading", { name: /Account Details/i })
+  ).toBeVisible();
+
+  await expect(
+    page.getByRole("heading", { name: /My Account/i })
+  ).toBeVisible();
+
+  await expect(page.getByText(/Email/i)).toBeVisible();
+  await expect(page.getByText(REGISTERED_USER_CREDENTIALS.email)).toBeVisible();
 
   // Shop for items as registered user
-  await addProductToCart({page});
+  await page.goto(config.RETAIL_APP_HOME);
 
+  await page.getByRole("link", { name: "Womens" }).hover();
+  const topsNav = await page.getByRole("link", { name: "Tops", exact: true });
+  await expect(topsNav).toBeVisible();
+
+  await topsNav.click();
+
+  // PLP
+  const productTile = page.getByRole("link", {
+    name: /Cotton Turtleneck Sweater/i,
+  });
+  // selecting swatch
+  const productTileImg = productTile.locator("img");
+  await productTileImg.waitFor({state: 'visible'})
+  const initialSrc = await productTileImg.getAttribute("src");
+  await expect(productTile.getByText(/From \$39\.99/i)).toBeVisible();
+
+  await productTile.getByLabel(/Black/, { exact: true }).hover();
+  // Make sure the image src has changed
+  await expect(async () => {
+    const newSrc = await productTileImg.getAttribute("src")
+    expect(newSrc).not.toBe(initialSrc)
+  }).toPass()
+  await expect(productTile.getByText(/From \$39\.99/i)).toBeVisible();
+  await productTile.click();
+
+  // PDP
+  await expect(
+    page.getByRole("heading", { name: /Cotton Turtleneck Sweater/i })
+  ).toBeVisible();
+  await page.getByRole("radio", { name: "L", exact: true }).click();
+
+  await page.locator("button[data-testid='quantity-increment']").click();
+
+  // // Selected Size and Color texts are broken into multiple elements on the page.
+  // // So we need to look at the page URL to verify selected variants
+  const updatedPageURL = await page.url();
+  const params = updatedPageURL.split("?")[1];
+  expect(params).toMatch(/size=9LG/i);
+  expect(params).toMatch(/color=JJ169XX/i);
+  await page.getByRole("button", { name: /Add to Cart/i }).click();
+
+  const addedToCartModal = page.getByText(/2 items added to cart/i);
+
+  await addedToCartModal.waitFor();
+
+  await page.getByLabel("Close").click();
   // cart
   await page.getByLabel(/My cart/i).click();
 
@@ -86,7 +146,7 @@ test("Registered shopper can checkout items", async ({ page }) => {
   await expect(
     page.getByRole("heading", { name: /Shipping & Gift Options/i })
   ).toBeVisible();
-  await page.waitForLoadState();
+  await page.waitForTimeout(2000);
 
   const continueToPayment = page.getByRole("button", {
     name: /Continue to Payment/i,
@@ -133,35 +193,4 @@ test("Registered shopper can checkout items", async ({ page }) => {
   await expect(
     page.getByRole("link", { name: /Cotton Turtleneck Sweater/i })
   ).toBeVisible();
-
-  // order history
-  await validateOrderHistory({page});
-});
-
-/**
- * Test that registered shoppers can navigate to PDP and add a product to wishlist
- */
-test("Registered shopper can add item to wishlist", async ({ page }) => {
-  const isLoggedIn = await loginShopper({
-    page,
-    userCredentials: REGISTERED_USER_CREDENTIALS
-  })
-
-  if(!isLoggedIn) {
-    await registerShopper({page, userCredentials: REGISTERED_USER_CREDENTIALS})
-  }
-
-  // Navigate to PDP
-  await navigateToPDPDesktop({page});
-
-  // add product to wishlist
-  await expect(
-    page.getByRole("heading", { name: /Cotton Turtleneck Sweater/i })
-  ).toBeVisible();
-
-  await page.getByRole("radio", { name: "L", exact: true }).click();
-  await page.getByRole("button", { name: /Add to Wishlist/i }).click()
-
-  // wishlist
-  await validateWishlist({page})
 });