+++ packages/commerce-sdk-react/src/hooks/ShopperLogin/cache.ts
@@ -15,12 +15,10 @@ export const cacheUpdateMatrix: CacheUpdateMatrix<ApiClients['shopperLogin']> = 
             remove: [{queryKey: ['/commerce-sdk-react']}]
         }
     },
-    authorizeCustomer: noop,
     getAccessToken: noop,
     getSessionBridgeAccessToken: noop,
     getTrustedSystemAccessToken: noop,
     getTrustedAgentAccessToken: noop,
-    getPasswordResetToken: noop,
     resetPassword: noop,
     getPasswordLessAccessToken: noop,
     revokeToken: noop,