+++ packages/template-retail-react-app/app/components/header/index.test.js
@@ -64,7 +64,7 @@ test('renders Header', async () => {
     await waitFor(() => {
         const menu = screen.getByLabelText('Menu')
         const logo = screen.getByLabelText('Logo')
-        const account = screen.getByLabelText('My Account')
+        const account = screen.getByLabelText('My account')
         const cart = screen.getByLabelText('My cart, number of items: 0')
         const wishlist = screen.getByLabelText('Wishlist')
         const searchInput = document.querySelector('input[type="search"]')
@@ -93,7 +93,7 @@ test('renders Header with event handlers', async () => {
     await waitFor(() => {
         const menu = screen.getByLabelText('Menu')
         const logo = screen.getByLabelText('Logo')
-        const account = screen.getByLabelText('My Account')
+        const account = screen.getByLabelText('My account')
         const cart = screen.getByLabelText('My cart, number of items: 0')
         expect(menu).toBeInTheDocument()
         fireEvent.click(menu)
@@ -157,7 +157,7 @@ test('route to account page when an authenticated users click on account icon', 
         const accountTrigger = screen.getByLabelText('Open account menu')
         expect(accountTrigger).toBeInTheDocument()
     })
-    const accountIcon = screen.getByLabelText('My Account')
+    const accountIcon = screen.getByLabelText('My account')
     fireEvent.click(accountIcon)
     await waitFor(() => {
         expect(history.push).toHaveBeenCalledWith(createPathWithDefaults('/account'))
@@ -208,7 +208,7 @@ test('shows dropdown menu when an authenticated users hover on the account icon'
         const accountTrigger = screen.getByLabelText('Open account menu')
         expect(accountTrigger).toBeInTheDocument()
     })
-    const accountIcon = screen.getByLabelText('My Account')
+    const accountIcon = screen.getByLabelText('My account')
     fireEvent.click(accountIcon)
     await waitFor(() => {
         expect(history.push).toHaveBeenCalledWith(createPathWithDefaults('/account'))