+++ packages/template-retail-react-app/app/pages/product-list/index.jsx
@@ -25,7 +25,6 @@ import {
     SimpleGrid,
     Grid,
     Select,
-    Heading,
     Text,
     FormControl,
     Stack,
@@ -450,7 +449,6 @@ const ProductList = (props) => {
                         </Box>
                     </Stack>
 
-                    {/* Filter Button for Mobile */}
                     <HideOnDesktop>
                         <Stack spacing={6}>
                             <PageHeader
@@ -636,12 +634,12 @@ const ProductList = (props) => {
                 <ModalOverlay />
                 <ModalContent top={0} marginTop={0}>
                     <ModalHeader>
-                        <Heading as="h1" fontWeight="bold" fontSize="2xl">
+                        <Text fontWeight="bold" fontSize="2xl">
                             <FormattedMessage
                                 defaultMessage="Filter"
                                 id="product_list.modal.title.filter"
                             />
-                        </Heading>
+                        </Text>
                     </ModalHeader>
                     <ModalCloseButton />
                     <ModalBody py={4}>