+++ packages/template-retail-react-app/app/utils/locale.js
@@ -14,11 +14,10 @@ import fetch from 'cross-fetch'
 /**
  * Dynamically import the translations/messages for a given locale
  * @param {string} locale
- * @param {string} origin
  * @returns {Promise<Object>} The messages (compiled in AST format) in the given locale.
  *      If the translation file is not found, return an empty object (so react-intl would fall back to the inline messages)
  */
-export const fetchTranslations = async (locale, origin) => {
+export const fetchTranslations = async (locale) => {
     const targetLocale =
         typeof window === 'undefined'
             ? process.env.USE_PSEUDOLOCALE === 'true'
@@ -27,7 +26,7 @@ export const fetchTranslations = async (locale, origin) => {
             : locale
 
     try {
-        const file = `${origin || getAppOrigin()}${getAssetUrl(
+        const file = `${getAppOrigin()}${getAssetUrl(
             `static/translations/compiled/${targetLocale}.json`
         )}`
         const response = await fetch(file)